Handle empty object reads
If a storage implementation returns an EncodedObject with size 0 and it
happens to return a bytes.Reader from the object's Reader method,
instead of, for example, a bytes.Buffer (as memory storage does, which
quietly skirts around this problem), then Parser.get will receive an
io.EOF and fail. This is kind of an odd case because the memory storage
implementation doesn't hit it by using bytes.Buffer, which has some
unusual behavior with EOF on empty buffers.

bytes.Buffer, on the other hand, will immediately return io.EOF if it
has an empty slice to read from. The same is true of strings.Reader
(they're just about the same), which will also cause Parser.get to
fail.

To work around this, if the encoded object size is 0 and an io.EOF is
returned, discard the EOF and proceed as usual.

Signed-off-by: Noel Cower <ncower@nil.dev>